apps/emoji/requirements.txt
jinja2
requests
pyyaml
click
